Use double instead size_t for trensferred/read, so that lb doesn't break on trnasferred mode when 2G of data has been send/read.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@300816 13f79535-47bb-0310-9956-ffa450edef68

+++ jk/native/common/jk_lb_worker.c
@@ -225,8 +225,8 @@ static worker_record_t *find_best_bydomain(lb_worker_t *p,
 {
     unsigned int i;
     int total_factor = 0;
-    size_t mytraffic = 0;
-    size_t curmin = 0;
+    double mytraffic = 0;
+    double curmin = 0;
 
     worker_record_t *candidate = NULL;
     if (p->lblock == JK_LB_LOCK_PESSIMISTIC)
@@ -313,8 +313,8 @@ static worker_record_t *find_best_bytraffic(lb_worker_t *p,
                                              jk_logger_t *l)
 {
     unsigned int i;
-    size_t mytraffic = 0;
-    size_t curmin = 0;
+    double mytraffic = 0;
+    double curmin = 0;
     worker_record_t *candidate = NULL;
 
     if (p->lblock == JK_LB_LOCK_PESSIMISTIC)
